move stack out to be reusable
