+++ src/compiler/utils/FrameStack.js
@@ -1,26 +1,4 @@
-let Stack = function() {
-  let history = [],
-      memory = [];
-  let count = 0;
-  function current() {
-    return history.length ? history[history.length - 1] : null;
-  }
-  this.current = current;
-  this.enter = function(item) {
-    history.push(item || count++);
-  };
-  this.leave = function() {
-    history.pop();
-  };
-  this.jump = function() {
-    memory.push(history);
-    history = [];
-  };
-  this.drop = function() {
-    history = memory.pop();
-  };
-  return this;
-};
+import Stack from './Stack';
 
 let FrameStack = function() {
   let tdStack = new Stack();
@@ -71,4 +49,4 @@ let FrameStack = function() {
 };
 
 
-module.exports = FrameStack;
+export default FrameStack;
